Move StorageInstanceId to mz-catalog-types
Move the StorageInstanceId type from mz-storage-types to mz-catalog-types.
The original crate now re-exports the type for backwards compatibility.
